python-embedding-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.7.0
qdrant-client==1.6.9
pydantic==2.5.0
python-multipart==0.0.6
PyMySQL==1.1.0
python-dotenv==1.0.0
